Improve the performance of running execution
Change the way a state that needs to be run is treated.

Every state that needed to be ran was stored is a redis queue, which
is good for a multi cluster setup, since it ensutre that every node
will naturaly have a comparable workload, but it add a lot of overhead.

Now, the next state of a currently running state will be immediatly
ran in the same node, since most state are extremely quick to run.
First state of an execution, as well as Task state when they finish
and wait state are still gonna be put in queues however.