Add support for stopping execution in wait state
This is done by veryfing if the execution is still running when we
manage a failed or success state (so during the end of the state
processing).

Add test that stops an execution during the waiting time of a Wait
state. To do that, I added  the fields 'stopAfterSeconds' that trigger
a stopExecution X seconds after the execution started.