Improve stopExecution while a Task is running
Manage and test an execution being stopped when a task state is launched,
but the task was not picked up by a worker and the execution stops.
In that case, we do not remove the task from the activiy queue in redis,
but instead, each time a task in retrieved in GetActivityTask, we
check if the execution is still running.