Improve Parallel state management
Parallel state now register in the currentRunningState Set in Redis.
When the parallel state finishes, it removes it from the set. this set
is used to know which parallel state is running, so when a state is
stopped, we can iterate over this set and remove the info of the
running state from redis.