Modify Running WaitState management
Running wait state are stored in redis with their token as key. this is
this token that is stored in the sorted set used as timer.

This change was made in order to be more consistent, and also because,
in order to match AWS Behavior, we needed to retrieve the wait state
info if a branche of a parallel state failed.